Added TicketScan scene and implemented QR code scanning in OP_QR_Scan.cs with scene transitions.
TicketScan Scene: Introduced a new scene, TicketScan, which is intended for the first-time scan of the user's ticket.

QR Code Scanning: The OP_QR_Scan.cs script successfully scans the QR code, stores its information in PlayerPrefs, and transitions to the Main scene.

Scene Transition Delay: Added a delay in scene transitions to ensure successful loading. This is a bandaid but works for now.

Known issues:

Camera Orientation: Need to adjust the camera to be either vertical or horizontal based on how the phone is held, along with ensuring the correct aspect ratio and rotation.

Permission Handling: Currently, the first run of the app asks for camera permission but doesn't update the camera feed once permission is granted, requiring a manual app reload.